SingleCellExperiment object to Seurat object

Transfer SingleCellExperiment object to a Seurat object for preparation for DR.SC model fitting; see our [DR.SC package website](https://feiyoung.github.io/DR.SC/index.html) for more usage of DR.SC.

sce2seurat(sce, verbose= TRUE)

Arguments

sce

a SingleCellExperiment object, at least including the raw gene count expression matrix.

verbose

an optional logical value, whether output the information.

Value

Return a Seurat object.

Author

Wei Liu

Note

nothing

See also

Examples

  library(SingleCellExperiment)
  dir <- system.file(
    file.path("extdata", "10xVisium", "section1"),
    package = "SpatialExperiment")

  # read in counts
  fnm <- file.path(dir, "raw_feature_bc_matrix")
  sce <- DropletUtils::read10xCounts(fnm)
  colnames(sce) <- paste0("cell", 1:ncol(spe))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'ncol': object 'spe' not found

  seu <- sce2seurat(sce)
#> Transfer SingleCellExperiment object to a Seurat object
#> preparation for  model fitting
#> Warning: Some cells in meta.data not present in provided counts matrix
#> Error in CreateAssayObject(counts = counts, min.cells = min.cells, min.features = min.features,     row.names = row.names): No cell names (colnames) names present in the input matrix
  head(seu)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'seu' not found

  ## Fit DR-SC model
  library(DR.SC)
#> Loading required package: spatstat.geom
#> Warning: package 'spatstat.geom' was built under R version 4.1.3
#> Loading required package: spatstat.data
#> Warning: package 'spatstat.data' was built under R version 4.1.3
#> spatstat.geom 2.4-0
#> 
#> Attaching package: 'spatstat.geom'
#> The following object is masked from 'package:SummarizedExperiment':
#> 
#>     shift
#> The following object is masked from 'package:GenomicRanges':
#> 
#>     shift
#> The following objects are masked from 'package:IRanges':
#> 
#>     reflect, shift
#> DR.SC :  Joint dimension reduction and spatial clustering is conducted for
#> Single-cell RNA sequencing and spatial transcriptomics data, and more details can be referred to
#> Wei Liu, Xu Liao, Yi Yang, Huazhen Lin, Joe Yeong, Xiang Zhou, Xingjie Shi and Jin Liu. (2022) <doi:10.1093/nar/gkac219>. It is not only computationally efficient and scalable to the sample size increment, but also is capable of choosing the smoothness parameter and the number of clusters as well.
#> 
#> Attaching package: 'DR.SC'
#> The following object is masked from 'package:PRECAST':
#> 
#>     selectModel
  seu <- NormalizeData(seu)
#> Error in NormalizeData(seu): object 'seu' not found
  seu <- FindVariableFeatures(seu)
#> Error in FindVariableFeatures(seu): object 'seu' not found
  seu <- DR.SC(seu = seu, K=4, platform = "scRNAseq")
#> Error in DR.SC(seu = seu, K = 4, platform = "scRNAseq"): object 'seu' not found